refactor: moving mocked user and organization to commoon services
