+++ apps/bilan-carbone/src/services/permissions/organization.test.ts
@@ -2,11 +2,12 @@ import * as dbAccount from '@/db/account'
 import * as dbOrganization from '@/db/organization'
 import * as dbStudy from '@/db/study'
 import * as dbUser from '@/db/user'
-import { mockedOrganizationId, mockedOrganizationVersionId } from '@/tests/utils/models/organization'
+import { mockedOrganizationVersionId } from '@/tests/utils/models/organization'
 import { getMockedAuthUser } from '@/tests/utils/models/user'
 import * as organizationUtils from '@/utils/organization'
 import * as userUtils from '@/utils/user'
 import { Role } from '@abc-transitionbascarbone/db-common/enums'
+import { mockedOrganizationId } from '@abc-transitionbascarbone/services/tests/models/organization'
 import { expect } from '@jest/globals'
 import { UserSession } from 'next-auth'
 import * as authModule from '../auth'
