+++ apps/bilan-carbone/src/services/serverFunctions/user.test.ts
@@ -18,10 +18,12 @@ import {
 import { addSite } from '@/db/site'
 import { addUser, getUserByEmail, organizationVersionActiveAccountsCount, updateAccount, validateUser } from '@/db/user'
 import { EMAIL_SENT, NOT_ASSOCIATION_SIRET, REQUEST_SENT, UNKNOWN_SIRET_OR_CNC } from '@/services/permissions/check'
-import { mockedOrganizationId, mockedOrganizationVersionId } from '@/tests/utils/models/organization'
-import { mockedAccountId, mockedUserId } from '@/tests/utils/models/user'
+import { mockedOrganizationVersionId } from '@/tests/utils/models/organization'
+import { mockedAccountId } from '@/tests/utils/models/user'
 import { sendActivationRequest } from '@abc-transitionbascarbone/services/email/email'
 import { NOT_AUTHORIZED } from '@abc-transitionbascarbone/services/permissions/check'
+import { mockedOrganizationId } from '@abc-transitionbascarbone/services/tests/models/organization'
+import { mockedUserId } from '@abc-transitionbascarbone/services/tests/models/user'
 import { getCompanyName, getValidAssociationNameBySiret } from '../associationApi'
 import { getDeactivableFeatureRestrictions } from './deactivableFeatures'
 import { activateEmail, signUpWithSiretOrCNC } from './user'
