+++ apps/mip/src/tests/utils/models/user.ts
@@ -1,29 +1,11 @@
 import type { AccountMip, Prisma, User } from '@abc-transitionbascarbone/db-common'
-import { Environment, Level, Role, UserSource, UserStatus } from '@abc-transitionbascarbone/db-common/enums'
+import { Environment, Level, Role, UserStatus } from '@abc-transitionbascarbone/db-common/enums'
+import { mockedOrganizationId } from '@abc-transitionbascarbone/services/tests/models/organization'
+import { mockedDbUser, mockedUser, mockedUserId } from '@abc-transitionbascarbone/services/tests/models/user'
 import { Session, UserSession } from 'next-auth'
-import { mockedOrganizationId, mockedOrganizationVersionMipId } from './organization'
+import { mockedOrganizationVersionMipId } from './organization'
 
-export const mockedUserId = 'mocked-user-id'
 export const mockedAccountMipId = 'mocked-account-mip-id'
-
-export const mockedUser = {
-  id: mockedUserId,
-  email: 'mocked@email.com',
-  firstName: 'Mocked',
-  lastName: 'User',
-}
-
-export const mockedDbUser = {
-  ...mockedUser,
-  createdAt: new Date('2025-01-01T00:00:00.000Z'),
-  updatedAt: new Date('2025-01-01T00:00:00.000Z'),
-  importedFileDate: null,
-  password: null,
-  resetToken: null,
-  formationFormStartTime: null,
-  source: UserSource.CRON,
-  level: null,
-}
 
 const mockedAccountMip = {
   id: mockedAccountMipId,
