+++ packages/services/src/tests/models/organization.ts
@@ -1,0 +1,6 @@
+export const mockedOrganizationId = 'mocked-organization-id'
+
+export const mockedOrganization = {
+  id: mockedOrganizationId,
+  name: 'Mocked Organization',
+}
